# server/apidocs-info.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API referente ao Projeto Interdisciplinar - 5° Sem.",
        "description": "API contendo todas as funcionalidades/endpoints do PI.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "URL para testes dos endpoints"
        }
    ],
    "paths": {
        "/cadastro": {
            "post": {
                "summary": "Cadastro de usuário",
                "description": "Rota responsável pelo cadastro dos dados do usuário.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastroUsuario"
                            }
                        }
                    }
                }
            }
        },
        "/login":{
            "post": {
                "summary": "Login de usuário",
                "description": "Rota responsável por fazer o login do usuário.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUsuario"
                            }
                        }
                    }
                }
            }
        },
        "/user/:email":{
            "get": {
                "summary": "Dados do usuário",
                "description": "Rota responsável por retornar todos os dados do usuário.",
                "parameters": {
                    "in": "path",
                    "name": "email",
                    "schema": {
                        "type": "string"
                    },
                    "required": "true"
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CadastroUsuario": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            },
            "LoginUsuario": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            }
        }
    }
}